+++ tests/conftest.py
@@ -2,7 +2,7 @@ import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
 
-from complex_frogs.database.models import Base, ScrapedData, ScrapeTargets
+from src.database.models import Base, ScrapedData, ScrapeTargets
 
 from . import dummy_data as data
